pyproject.toml
[project]
name = "maptoposter"
version = "0.2.0"
description = "Generate beautiful, minimalist map posters for any city in the world."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ankur Gupta", email = "originalankur@github.com"}
]
keywords = ["map", "poster", "osm", "openstreetmap", "art", "visualization"]

dependencies = [
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fonttools==4.61.1",
    "geographiclib==2.1",
    "geopandas==1.1.2",
    "geopy==2.4.1",
    "idna==3.11",
    "kiwisolver==1.4.9",
    "lat-lon-parser==1.3.1",
    "matplotlib==3.10.8",
    "networkx==3.6.1",
    "numpy==2.4.0",
    "osmnx==2.0.7",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.1.0",
    "pyogrio==0.12.1",
    "pyparsing==3.3.1",
    "pyproj==3.7.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "replicate>=1.0.0",
    "requests==2.32.5",
    "scipy==1.16.3",
    "shapely==2.1.2",
    "six==1.17.0",
    "streamlit>=1.54.0",
    "tqdm==4.67.1",
    "tzdata==2025.3",
    "urllib3==2.6.3",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["create_map_poster"]
